+++ cmd/frontend/backend/repos_mock.go
@@ -11,6 +11,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
 	"github.com/sourcegraph/sourcegraph/pkg/api"
 	"github.com/sourcegraph/sourcegraph/pkg/errcode"
+	"github.com/sourcegraph/sourcegraph/pkg/gitserver"
 	"github.com/sourcegraph/sourcegraph/pkg/vcs/git"
 )
 
@@ -19,6 +20,7 @@ type MockRepos struct {
 	GetByName                 func(v0 context.Context, name api.RepoName) (*types.Repo, error)
 	AddGitHubDotComRepository func(name api.RepoName) error
 	List                      func(v0 context.Context, v1 db.ReposListOptions) ([]*types.Repo, error)
+	ListWithLongestInterval   func(v0 context.Context, v1 int) ([]string, error)
 	GetCommit                 func(v0 context.Context, repo *types.Repo, commitID api.CommitID) (*git.Commit, error)
 	ResolveRev                func(v0 context.Context, repo *types.Repo, rev string) (api.CommitID, error)
 	GetInventory              func(v0 context.Context, repo *types.Repo, commitID api.CommitID) (*inventory.Inventory, error)
@@ -104,7 +106,7 @@ func (s *MockRepos) MockResolveRev_NotFound(t *testing.T, wantRepo api.RepoID, w
 		if rev != wantRev {
 			t.Errorf("got rev %v, want %v", rev, wantRev)
 		}
-		return "", &git.RevisionNotFoundError{Repo: repo.Name, Spec: rev}
+		return "", &gitserver.RevisionNotFoundError{Repo: repo.Name, Spec: rev}
 	}
 	return
 }