+++ pkg/extsvc/github/repos.go
@@ -534,3 +534,24 @@ func (c *Client) ListRepositoriesForSearch(ctx context.Context, searchString str
 		HasNextPage: page*100 < response.TotalCount,
 	}, nil
 }
+
+type restTopicsResponse struct {
+	Names []string `json:"names"`
+}
+
+// ListTopicsOnRepository lists topics on the given repository.
+func (c *Client) ListTopicsOnRepository(ctx context.Context, ownerAndName string) ([]string, error) {
+	owner, name, err := SplitRepositoryNameWithOwner(ownerAndName)
+	if err != nil {
+		return nil, err
+	}
+
+	var result restTopicsResponse
+	if err := c.requestGet(ctx, "", fmt.Sprintf("/repos/%s/%s/topics", owner, name), &result); err != nil {
+		if HTTPErrorCode(err) == http.StatusNotFound {
+			return nil, ErrNotFound
+		}
+		return nil, err
+	}
+	return result.Names, nil
+}