+++ schema/schema.go
@@ -167,6 +167,7 @@ type BrandAssets struct {
 //
 // Only available in Sourcegraph Enterprise.
 type Branding struct {
+	BrandName         string       `json:"brandName,omitempty"`
 	Dark              *BrandAssets `json:"dark,omitempty"`
 	DisableSymbolSpin bool         `json:"disableSymbolSpin,omitempty"`
 	Favicon           string       `json:"favicon,omitempty"`
@@ -403,8 +404,9 @@ type OpenIDConnectAuthProvider struct {
 
 // OtherExternalServiceConnection description: Configuration for a Connection to Git repositories for which an external service integration isn't yet available.
 type OtherExternalServiceConnection struct {
-	Repos []string `json:"repos"`
-	Url   string   `json:"url,omitempty"`
+	Repos                 []string `json:"repos"`
+	RepositoryPathPattern string   `json:"repositoryPathPattern,omitempty"`
+	Url                   string   `json:"url,omitempty"`
 }
 
 // ParentSourcegraph description: URL to fetch unreachable repository details from. Defaults to "https://sourcegraph.com"
@@ -514,6 +516,8 @@ type SiteConfiguration struct {
 	GitMaxConcurrentClones            int                         `json:"gitMaxConcurrentClones,omitempty"`
 	GithubClientID                    string                      `json:"githubClientID,omitempty"`
 	GithubClientSecret                string                      `json:"githubClientSecret,omitempty"`
+	LsifUploadSecret                  string                      `json:"lsifUploadSecret,omitempty"`
+	LsifVerificationGithubToken       string                      `json:"lsifVerificationGithubToken,omitempty"`
 	MaxReposToSearch                  int                         `json:"maxReposToSearch,omitempty"`
 	ParentSourcegraph                 *ParentSourcegraph          `json:"parentSourcegraph,omitempty"`
 	RepoListUpdateInterval            int                         `json:"repoListUpdateInterval,omitempty"`