+++ web/src/enterprise/namespaces/routes.ts
@@ -1,0 +1,4 @@
+import { NamespaceAreaRoute } from '../../namespaces/NamespaceArea'
+import { namespaceAreaRoutes } from '../../namespaces/routes'
+
+export const enterpriseNamespaceAreaRoutes: ReadonlyArray<NamespaceAreaRoute> = [...namespaceAreaRoutes]