+++ CHANGELOG.md
@@ -15,9 +15,18 @@ All notable changes to Sourcegraph are documented in this file.
 
 ### Changed
 
+- Out-of-the-box TypeScript code intelligence is much better with an updated ctags version with a built-in TypeScript parser.
+
 ### Fixed
 
 ### Removed
+
+## 3.6.1 (unreleased)
+
+## Added
+
+- New site config option `branding.brandName` configures the brand name to display in the Sourcegraph \<title\> element.
+- `repositoryPathPattern` option added to the "Other" external service type for repository name customization.
 
 ## 3.6.0